v0.9.2 - новости на главной странице отображаются +- адекватно за счёт css - создание новости теперь проходит по ручке редактирования, убрал лишний эндпоинт - при сохранении новости в статус черновика или архива больше не требуется заполненное содержание - при запрете на редактирование/создание новости исходя из ошибки бэкэнда, будет происходить переадресация сразу на главную страницу

+++ web/js/editing_news/initButtons.js
@@ -17,7 +17,7 @@ function loadData(formData, btn) {
     }
 }
 
-function validateRequiredFields() {
+function validateRequiredFields(is_com) {
     const titleInput = document.getElementById("news-title");
     const commentInput = document.getElementById("news-comment");
 
@@ -26,7 +26,7 @@ function validateRequiredFields() {
         return false;
     }
 
-    if (!commentInput.value.trim()) {
+    if (is_com && !commentInput.value.trim()) {
         alert("Пожалуйста, заполните содержание новости.");
         return false;
     }
@@ -35,10 +35,11 @@ function validateRequiredFields() {
 }
 
 async function handleNewsAction(statusMessage, redirectStatus) {
-    if (!validateRequiredFields()) {
+    if (!validateRequiredFields(redirectStatus === 'ACTIVE')) {
         return;
     }
 
+    let new_id = -1
     const formData = new FormData();
     const btn = document.querySelector(`.action-btn[data-status="${redirectStatus}"]`);
     loadData(formData, btn);
@@ -56,8 +57,11 @@ async function handleNewsAction(statusMessage, redirectStatus) {
         const data = await response.json();
 
         if (data.success) {
+            new_id = data.id;
             alert(statusMessage);
-            window.location.href = `/news-list?status=${redirectStatus.toLowerCase()}`;
+            if (redirectStatus === 'ACTIVE')
+                window.location.href = `/`;
+            return new_id
         } else {
             alert("Ошибка: " + data.message);
         }
@@ -75,39 +79,56 @@ export function initButtons(newsID) {
     });
 
     document.getElementById("draft-btn").addEventListener("click", async function () {
-        await handleNewsAction("Новость успешно сохранена как черновик!", "DRAFT");
+        let new_id = await handleNewsAction("Новость успешно сохранена как черновик!", "DRAFT");
+        if (newsID === -1) {
+            window.location.href = `/editing_news?id=${new_id}`;
+        }
     });
 
     document.getElementById("archive-btn").addEventListener("click", async function () {
-        await handleNewsAction("Новость успешно архивирована!", "ARCHIVE");
-    });
-
-    document.getElementById("delete-btn").addEventListener("click", function (e) {
-        e.preventDefault();
-
-        const confirmed = window.confirm("Вы уверены, что хотите удалить эту новость? Отменить это действие будет невозможно");
-
-        if (confirmed) {
-            fetch(`/delete-news/${news_id}`, {
-                method: "DELETE",
-                credentials: "include",
-                headers: {
-                    "X-Requested-With": "XMLHttpRequest",
-                },
-            })
-                .then(response => {
-                    return response.json().then(data => {
-                        if (!response.ok) {
-                            throw new Error(data.message || "Неизвестная ошибка");
-                        }
-                        alert("Новость удалена!");
-                        window.location.href = `/`;
-                    });
-                })
-                .catch(error => {
-                    console.error("Ошибка:", error);
-                    alert("Не удалось удалить новость: " + error.message);
-                });
+        let new_id = await handleNewsAction("Новость успешно архивирована!", "ARCHIVE");
+        if (newsID === -1) {
+            window.location.href = `/editing_news?id=${new_id}`;
         }
     });
+
+    const delete_btn = document.getElementById("delete-btn")
+
+    if (newsID !== -1) {
+        delete_btn.innerHTML = 'Удалить'
+        delete_btn.addEventListener("click", function (e) {
+            e.preventDefault();
+
+            const confirmed = window.confirm("Вы уверены, что хотите удалить эту новость? Отменить это действие будет невозможно");
+
+            if (confirmed) {
+                fetch(`/delete-news/${newsID}`, {
+                    method: "DELETE",
+                    credentials: "include",
+                    headers: {
+                        "X-Requested-With": "XMLHttpRequest",
+                    },
+                })
+                    .then(response => {
+                        return response.json().then(data => {
+                            if (!response.ok) {
+                                throw new Error(data.message || "Неизвестная ошибка");
+                            }
+                            alert("Новость удалена!");
+                            window.location.href = `/`;
+                        });
+                    })
+                    .catch(error => {
+                        console.error("Ошибка:", error);
+                        alert("Не удалось удалить новость: " + error.message);
+                    });
+            }
+        });
+    } else {
+        delete_btn.innerHTML = 'Отменить создание'
+        delete_btn.addEventListener("click", function (e) {
+            e.preventDefault();
+            window.location.href = `/`;
+        });
+    }
 }
